ingredientsをcomponentに切り出す
