+++ frontend/src/features/RegisterShop/Ingredients/index.graphql
@@ -1,7 +1,6 @@
-query Ingredients {
+query RegisterShopIngredients {
   ingredients {
     id
     name
-    category
   }
 }
