+++ frontend/src/features/RegisterShop/index.tsx
@@ -1,35 +1,21 @@
-import { useIngredientsQuery } from '@/graphql/generated';
-import { useCustomOptions } from '@/hooks/useOptions';
 import {
   Button,
-  Checkbox,
-  CheckboxGroup,
   Flex,
   FormControl,
   FormErrorMessage,
   FormHelperText,
   FormLabel,
   Input,
-  SkeletonText,
 } from '@chakra-ui/react';
 import { memo, useState } from 'react';
-import { Controller, useFieldArray, useForm } from 'react-hook-form';
+import { useFieldArray, useForm } from 'react-hook-form';
 import Genres from './Genres';
+import Ingredients from './Ingredients';
 import { useHandler } from './hooks';
 import { FormValues, PreviewType } from './index.d';
 import styles from './index.module.scss';
 
 const RegisterShop = memo(() => {
-  const { data, loading } = useIngredientsQuery();
-
-  const ingredients = data?.ingredients || [];
-
-  const options = useCustomOptions({
-    items: ingredients,
-    getLabel: (item) => item.name,
-    ignoreDefault: true,
-  });
-
   const {
     register,
     handleSubmit,
@@ -146,47 +132,7 @@ const RegisterShop = memo(() => {
               </FormControl>
             </Flex>
             <Flex mb={6}>
-              <FormControl
-                isInvalid={
-                  errors.menus && !!errors.menus[index]?.ingredients?.message
-                }
-              >
-                <Flex mb={2}>
-                  <FormLabel className={styles.label}>アレルギー情報</FormLabel>
-                </Flex>
-                <SkeletonText
-                  isLoaded={!loading}
-                  skeletonHeight={4}
-                  spacing='5'
-                >
-                  <Controller
-                    name={`menus.${index}.ingredients`}
-                    control={control}
-                    render={({ field: { ref, ...rest } }) => (
-                      <CheckboxGroup {...rest}>
-                        <Flex gap={4} wrap='wrap'>
-                          {options.map((option) => (
-                            <Checkbox value={option.value} key={option.key}>
-                              {option.label}
-                            </Checkbox>
-                          ))}
-                        </Flex>
-                      </CheckboxGroup>
-                    )}
-                    rules={{
-                      validate: {
-                        atLeastOneRequired: (value) =>
-                          (value && value.length >= 1) ||
-                          '1つ以上選択してください',
-                      },
-                    }}
-                  />
-                </SkeletonText>
-                <FormErrorMessage>
-                  {errors.menus &&
-                    errors.menus[index]?.ingredients?.message?.toString()}
-                </FormErrorMessage>
-              </FormControl>
+              <Ingredients errors={errors} control={control} index={index} />
             </Flex>
             <Flex mb={8}>
               <FormControl>
